fix(pvpve): 第二轮完备性 critic 3 漏网缺陷全修(1 HIGH 漂移失效 + 2 单源)
C1(HIGH) diplomacyCore.clampFavorability 每 tick Math.round→亚 0.5/tick 漂移全丢、
被动漂移子系统失效(违 doc-04 累加后 clamp)；去 round 让 favorability 连续浮点累积
+ faction_diplomacy.favorability 列 INTEGER→REAL。
C2(MED) FactionDiplomacyService.advanceEdge nemesisStreak 只写(a,b)单向却驱动对称态
→交替调用两向各自计数永不达阈；镜像写两行(单源)。
C3(LOW) FactionRepository.upsertFaction 部分写清 createdAt→存在则保留守卫。
第一轮 4 修复经 workflow 二次对抗验证全 CORRECT_AND_COMPLETE。+3 回归。全绿 2103 测试。

+++ backend/repositories/FactionDiplomacyRepository.js
@@ -31,7 +31,7 @@ class FactionDiplomacyRepository {
         worldId TEXT,
         fromFactionId TEXT,
         toFactionId TEXT,
-        favorability INTEGER DEFAULT 0,
+        favorability REAL DEFAULT 0,
         state TEXT DEFAULT 'neutral',
         nemesisStreak INTEGER DEFAULT 0,
         since TEXT,
